+++ core/selection.js
@@ -2019,7 +2019,9 @@
 			// Handle special case - fake selection of table cells.
 			if ( editor && editor.plugins.tableselection &&
 				CKEDITOR.plugins.tableselection.isSupportedEnvironment &&
-				isTableSelection( ranges ) && !isSelectingTable
+				isTableSelection( ranges ) &&
+				!isSelectingTable &&
+				!ranges[ 0 ]._getTableElement( { table: 1 } ).hasAttribute( 'data-cke-tableselection-ignored' )
 			) {
 				performFakeTableSelection.call( this, ranges );
 				return;